feat: add image update asociation to product model

+++ src/controllers/image-controller.js
@@ -1,7 +1,11 @@
 const { Image, Product } = require("../db");
-const { uploadImageToCloudinary, deleteImageFromCloudinary, createImageInDataBase } = require("../utils/image-util");
+const {
+  uploadImageToCloudinary,
+  deleteImageFromCloudinary,
+  createImageInDataBase,
+} = require("../utils/image-util");
 
-const uploadImagesController = async (id, files) => {
+const uploadImages = async (id, files) => {
   try {
     const product = await Product.findByPk(id);
     if (!product) throw new Error(`Product with ID ${id} not found.`);
@@ -9,7 +13,6 @@ const uploadImagesController = async (id, files) => {
     const uploadedImages = await Promise.all(uploadPromises);
     const createPromises = uploadedImages.map((image) => createImageInDataBase(image));
     const imageInstances = await Promise.all(createPromises);
-    await product.addImages(imageInstances);
     return imageInstances;
   } catch (error) {
     console.error("Error uploading images:", error);
@@ -17,14 +20,15 @@ const uploadImagesController = async (id, files) => {
   }
 };
 
-const deleteImagesController = async (images) => {
+const deleteImages = async (imagesToDelete) => {
   try {
-    await Promise.all(
-      images.map(async (image) => {
-        await deleteImageFromCloudinary(image.publicId);
-        await Image.destroy({ where: { id: image.id } });
-      })
-    );
+    if (imagesToDelete.length)
+      await Promise.all(
+        imagesToDelete.map(async (image) => {
+          await deleteImageFromCloudinary(image.publicId);
+          await Image.destroy({ where: { id: image.id } });
+        })
+      );
     return "Images deleted successfully";
   } catch (error) {
     console.error("Error deleting image:", error);
@@ -33,6 +37,6 @@ const deleteImagesController = async (images) => {
 };
 
 module.exports = {
-  uploadImagesController,
-  deleteImagesController,
+  uploadImages,
+  deleteImages,
 };
